+++ changelog.md
@@ -3,3 +3,5 @@
 ## 2021-01-11 
 
 - Modified sample `dockerstarter.yaml`
+- Added (optional) configuration `image` to a container configuration to specify the image name
+- Added facilities to check if an image exists and to `docker pull` it if missing